MediBotAINew-main/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
ollama==0.1.7
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.0
python-multipart==0.0.6
reportlab==4.0.7
google-generativeai==0.3.2
supabase==2.3.0
python-dotenv==1.0.0